pyproject.toml
[project]
name = "evals-chat-with-dashboard"
version = "0.1.0"
description = "Accuracy evaluation harness for Dalgo's chat-with-dashboard feature"
requires-python = ">=3.10"
dependencies = [
    "websockets>=13.0",
    "python-dotenv>=1.0",
    "deepeval>=2.0",
    "httpx>=0.27",
]
